Add endGame logic and img
