+++ client/src/components/QUIZ_GAME/EndGame/EndGame.module.css
@@ -1,0 +1,21 @@
+.endGame {
+  width: 300px;
+  height: 180px;
+  margin: 10px auto;
+  border: 1px solid black;
+  text-align: center;
+  padding: 5px;
+  background-image: url('https://p4.wallpaperbetter.com/wallpaper/606/607/750/the-witcher-the-witcher-3-wild-hunt-logo-wallpaper-preview.jpg');
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.title {
+  color: white;
+  text-align: center;
+  font-size: 24px;
+}
