+++ client/src/components/QUIZ_GAME/mock.json
@@ -5,7 +5,8 @@
     "var1": "Вариант ответа 1",
     "var2": "Вариант ответа 2",
     "current": "Вариант ответа 1",
-    "status": null
+    "status": null,
+    "img": "https://cdnb.artstation.com/p/assets/images/images/008/185/067/medium/stefan-s-untitled-1.jpg?1511032697"
   },
   {
     "id": 2,
@@ -13,7 +14,8 @@
     "var1": "Вариант ответа 1",
     "var2": "Вариант ответа 2",
     "current": "Вариант ответа 2",
-    "status": null
+    "status": null,
+    "img": "https://cdnb.artstation.com/p/assets/images/images/008/185/069/medium/stefan-s-fight-6.jpg?1511032700"
   },
   {
     "id": 3,
@@ -21,7 +23,8 @@
     "var1": "Вариант ответа 1",
     "var2": "Вариант ответа 2",
     "current": "Вариант ответа 2",
-    "status": null
+    "status": null,
+    "img": "https://cdnb.artstation.com/p/assets/images/images/008/185/097/medium/stefan-s-fight-3.jpg?1511032707"
   },
   {
     "id": 4,
@@ -29,7 +32,8 @@
     "var1": "Вариант ответа 1",
     "var2": "Вариант ответа 2",
     "current": "Вариант ответа 1",
-    "status": null
+    "status": null,
+    "img": "https://cdna.artstation.com/p/assets/images/images/008/185/148/medium/stefan-s-vizima-2.jpg?1511032787"
   },
   {
     "id": 5,
@@ -37,6 +41,7 @@
     "var1": "Вариант ответа 1",
     "var2": "Вариант ответа 2",
     "current": "Вариант ответа 1",
-    "status": null
+    "status": null,
+    "img": "https://cdnb.artstation.com/p/assets/images/images/008/185/153/medium/stefan-s-swamp-2.jpg?1511032788"
   }
 ]
